% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusbootglm.R
\name{clusbootglm}
\alias{clusbootglm}
\title{Fit (generalized) linear models with the cluster bootstrap}
\usage{
clusbootglm(model, data, clusterid, family = gaussian, B = 5000,
  confint.level = 0.95, no_cores = 1)
}
\arguments{
\item{model}{(generalized) linear model to be fitted with the cluster bootstrap.}

\item{data}{dataframe that contains the data.}

\item{clusterid}{variable in data that identifies the clusters.}

\item{family}{currently, only Gaussian is supported.}

\item{B}{number of bootstrap samples.}

\item{confint.level}{level of confidence interval.}

\item{no_cores}{number of CPU cores to be used.}
}
\value{
\code{clusbootglm} produces an object of class \code{"clusboot"}, containing the following relevant components:
\item{coefficients}{A matrix of \code{B} rows, containing the parameter estimates for all bootstrap samples.}
\item{lm.coefs}{Parameter estimates from a single (generalized) linear model.}
\item{boot.coefs}{Mean values of the paramater estimates, derived from the bootstrap coefficients.}
\item{boot.sds}{Standard deviations of cluster bootstrap parameter estimates.}
\item{ci.level}{User defined confidence interval level.}
\item{percentile.interval}{Confidence interval based on percentiles, given the user defined confidence interval level.}
\item{parametric.interval}{Confidence interval based on \code{lm.coefs} and column standard deviations of \code{coefficients}, given the user defined confidence interval level.}
\item{BCa.interval}{Confidence interval based on percentiles with bias correction and acceleration, given the user defined confidence interval level.}
}
\description{
Perform a (generalized) linear model with the cluster bootstrap for analysis of clustered data.
}
\details{
Some useful methods for the obtained \code{clusboot} class object are \code{\link{summary.clusboot}}, \code{\link{coef.clusboot}} and \code{\link{plot.clusboot}}.
}
\examples{
\dontrun{
data(opposites)
clusbootglm(SCORE~Time*COG,data=opposites,clusterid=opposites$Subject)}
}
\author{
Mathijs Deen, Mark de Rooij
}
