% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_exports.R
\name{clusbootmatrix}
\alias{clusbootmatrix}
\title{Extract matrix with bootstrap samples}
\usage{
clusbootmatrix(object, whichsample = "all")
}
\arguments{
\item{object}{object of class \code{clusbootglm}, created with the \code{clusbootglm} function.}

\item{whichsample}{"all" for the complete bootstrap matrix, "failed" for bootstrap samples that returned NAs, or a vector of values for specific bootstrap samples.}
}
\description{
Obtain a matrix containing the \code{clusterid} values for the bootstrap samples in a \code{clusbootglm} object.
}
\examples{
\dontrun{
data(opposites)
cbglm.1 <- clusbootglm(SCORE~Time*COG,data=opposites,clusterid=opposites$Subject)
clusbootmatrix(cbglm.1, whichsample=c(1:5))}
}
\author{
Mathijs Deen
}
