% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_exports.R
\name{clusbootsample}
\alias{clusbootsample}
\title{Create data for specified bootstrap sample}
\usage{
clusbootsample(object, samplenr)
}
\arguments{
\item{object}{object of class \code{clusbootglm}, created with the \code{clusbootglm} function.}

\item{samplenr}{sample number for which the data frame should be created}
}
\description{
Returns the full data frame for a specified bootstrap sample
}
\examples{
\dontrun{
data(opposites)
cbglm.1 <- clusbootglm(SCORE~Time*COG,data=opposites,clusterid=opposites$Subject)
clusbootsample(cbglm.1, samplenr=1)}
}
\author{
Mathijs Deen, Mark de Rooij
}
