% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterBootstrap.R
\name{clusterBootstrap}
\alias{clusterBootstrap}
\title{Cluster Bootstrap}
\usage{
clusterBootstrap(
  df,
  clusters,
  replace,
  statFun = NULL,
  B = 5000,
  oob = FALSE,
  keepIndices = FALSE,
  nCores = 1L,
  ...
)
}
\arguments{
\item{df}{A data frame. The original dataset.}

\item{clusters}{A character vector of variable names that define the nested
structure of the data, ordered from highest to lowest level.}

\item{replace}{A logical vector indicating whether sampling should be with
replacement at each level. Should be of the same length as \code{clusters}.}

\item{statFun}{A function that takes a data frame (a bootstrap sample) and possibly an out-of-bag set,
and returns a numeric vector of statistics.}

\item{B}{Integer. The number of bootstrap samples to generate.}

\item{oob}{Indicates whether the out-of-bag set should be constructed besides the bootstrap sample. See 
details below.}

\item{keepIndices}{A logical value indicating whether the row numbers from \code{df} for the bootstrap 
sample should be returned. See Value element \code{indices}.}

\item{nCores}{The number of cores that are used for parallel computing.}

\item{...}{Additional arguments passed to \code{statFun}.}
}
\value{
\code{clusterBootstrap} returns an object of class \code{clusterBootstrap}, containing the 
following elements:
\item{call}{The function call}
\item{args}{Arguments passed to the function}
\item{estimates}{A list with the following elements:
  \itemize{
    \item \code{originalEstimates}: a \code{data.frame} with one row, 
    containing the return of \code{statFun} on the original data.
    \item \code{bootstrapEstimates}: a \code{data.frame} with B rows, 
    containing the return of \code{statFun} on each of the bootstrap samples.
    \item \code{bootstrapSE}: the bootstrap standard error(s) for all rows in \code{bootstrapEstimates}. 
  }
}
\item{indices}{If \code{keepIndices=TRUE}, a list of length \code{B} where each element contains the row numbers
of \code{df} that is in the corresponding bootstrap sample. If \code{keepIndices=FALSE}, this element is \code{NULL}.}
}
\description{
Performs bootstrapping on hierarchically structured data
using clustered or nested resampling at any level of the hierarchy. Allows bootstrapping of arbitrary
statistics computed from the resampled dataset.
}
\details{
When \code{oob = TRUE}, the out-of-bag (OOB) sample is determined for every bootstrap sample. The user 
defined \code{statFun} then takes two \code{data.frame}s: the first one is the bootstrap sample, the 
second one is the OOB sample. This makes it possible to do calculations on the OOB sample based on anything 
calculated from the bootstrap sample (e.g., calculate OOB prediction error).
}
\note{
The argument \code{stat_fun} got renamed softly to \code{statFun} in \code{ClusterBootstrap 2.1.0}. 
The old argument still works but will be fully deprecated in the future.
}
\examples{
\dontrun{
library(dplyr)
medData <- medication |>
filter(time \%\% 1 == 0, time < 4)
bootFun <- function(d) lm(pos ~ treat*time, data = d)$coefficients

# Resampling on the person level only
clusterBootstrap(df       = medData, 
                 clusters = "id", 
                 replace  = TRUE, 
                 statFun  = bootFun, 
                 B        = 5000)

# Resampling on the person level and the repeated measures level
clusterBootstrap(df       = medData, 
                 clusters = c("id", "time"), 
                 replace  = c(TRUE, TRUE), 
                 statFun  = bootFun, 
                 B        = 5000)

# Not resampling at one level 
# (e.g., by design all classes in a probed school are included, 
# but not all students in a class)
set.seed(2025)
n_school  <- 30
n_class   <- 8
n_student <- 15

demo <- expand.grid(
school  = paste0("S", 1:n_school),
class   = paste0("C", 1:n_class),
student = paste0("P", 1:n_student)) |>
  mutate(score1 = rnorm(n()),
         score2 = rnorm(n())) |>
  arrange(school, class, student) |>
  slice(1:(n() - 3)) # slightly unbalanced data
bootFun2 <- function(d) lm(score1 ~ score2, data = d)$coef
clusterBootstrap(df       = demo, 
                 clusters = c("school", "class", "student"),
                 replace  = c(TRUE, FALSE, TRUE),
                 statFun  = bootFun2,
                 B        = 1000)
}
}
\seealso{
\code{\link{clusterResample}} for the underlying resampling mechanism. \code{\link{confint.clusterBootstrap} 
for cluster bootstrap confidence intervals.}
}
\author{
Mathijs Deen
}
