% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusbootglm_s3_methods.R
\name{coef.clusboot}
\alias{coef.clusboot}
\title{Obtain coefficients from cluster bootstrap object}
\usage{
\method{coef}{clusboot}(object, type = "bootstrap", ...)
}
\arguments{
\item{object}{cluster bootstrap model.}

\item{type}{type of coefficient (bootstrap of GLM).}

\item{...}{other arguments.}
}
\description{
Obtain coefficients from cluster bootstrap object
}
\examples{
\dontrun{
data(opposites)
cbglm.1 <- clusbootglm(SCORE~Time*COG,data=opposites,clusterid=opposites$Subject)
coef(cbglm.1, type="bootstrap")}
}
\author{
Mathijs Deen
}
