% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.clusbootglm.R
\name{emmeans}
\alias{emmeans}
\title{Calculate estimated marginal means for a cluster bootstrap GLM}
\usage{
emmeans(object, confint.level = 0.95)
}
\arguments{
\item{object}{Object of class \code{clusbootglm}.}

\item{confint.level}{Level of the confidence interval. Should be in [0, 1]. Defaults to .95.}
}
\value{
\code{emmeans} returns estimated marginal means for a cluster bootstrap GLM. Does not work with \code{plot.clusbootemm} yet.
}
\description{
Returns estimated marginal means of a \code{clusbootglm} object. This is an experimental approach to the old \code{emm} function.
}
\examples{
\dontrun{
medication <- medication[medication$time \%\% 1 == 0,]
medication$time <- as.factor(medication$time)
set.seed(1)
model.1 <- clusbootglm(pos~time*treat, clusterid = id, data = medication)
emmeans(object = model.1)}
}
\author{
Mathijs Deen
}
