% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbag.R
\name{inbag}
\alias{inbag}
\alias{inbag.clusterBootstrap}
\title{Extract in-bag bootstrap sample}
\usage{
inbag(object, samplenr, ...)

\method{inbag}{clusterBootstrap}(object, samplenr, ...)
}
\arguments{
\item{object}{A bootstrap object}

\item{samplenr}{Integer. The bootstrap sample number to extract}

\item{...}{Additional arguments passed to methods}
}
\value{
A data frame containing the in-bag rows for the requested sample
}
\description{
Extracts the in-bag (bootstrap) sample for a given iteration from a
bootstrap object.
}
\author{
Mathijs Deen
}
