% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oob.R
\name{oob}
\alias{oob}
\alias{oob.clusterBootstrap}
\title{Extract out-of-bag sample}
\usage{
oob(object, samplenr, ...)

\method{oob}{clusterBootstrap}(object, samplenr, ...)
}
\arguments{
\item{object}{A bootstrap object}

\item{samplenr}{Integer. The bootstrap sample number to extract}

\item{...}{Additional arguments passed to methods}
}
\value{
A data frame containing the out-of-bag rows for the requested sample
}
\description{
Extracts the out-of-bag (OOB) sample for a given iteration from a
bootstrap object.
}
\author{
Mathijs Deen
}
