% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.clusbootemm.R
\name{pairs.clusbootemm}
\alias{pairs.clusbootemm}
\title{Calculate differences in estimated marginal means for a cluster bootstrap GLM}
\usage{
\method{pairs}{clusbootemm}(x, compare, at, compare.vals, confint.level = 0.95, ...)
}
\arguments{
\item{x}{Object of class \code{clusbootemm}.}

\item{compare}{Variable that distinguishes values at which to compare estimated marginal means (e.g., a group variable)}

\item{at}{Variable that distinguishes different levels at which to compare estimated marginal means (e.g., a time variable).}

\item{compare.vals}{Values of the \code{compare} variable that should be compared.}

\item{confint.level}{Level of the confidence interval.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
\code{pairs.clusbootemm} returns the differences between estimated marginal means for given values of \code{compare} at different levels of \code{at}.
}
\description{
Returns differences in estimated marginal means and their confidence intervals of a \code{clusbootglm} object.
}
\examples{
\dontrun{
medication <- medication[medication$time \%\% 1 == 0,]
medication$time <- as.factor(medication$time)
set.seed(1)
model.1 <- clusbootglm(pos~time*treat, clusterid = id, data = medication)
ems <- emmeans(object = model.1)
pairs(ems, compare=treat, at=time)}
}
\author{
Mathijs Deen
}
