% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairs.clusbootemm.R
\name{pairs}
\alias{pairs}
\title{bla}
\usage{
pairs(x, ...)
}
\arguments{
\item{x}{data to be used for the permutation test(s).}

\item{...}{additional arguments.}
}
\description{
blie
}
\author{
Mathijs Deen
}
