% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusbootglm_s3_methods.R
\name{plot.clusboot}
\alias{plot.clusboot}
\title{Plot confidence intervals of cluster bootstrap}
\usage{
\method{plot}{clusboot}(x, interval.type = "percentile",
  show.intercept = FALSE, ...)
}
\arguments{
\item{x}{object of class clusboot.}

\item{interval.type}{which confidence interval should be used. Choose \code{par} for parametric, \code{per} for percentile, or \code{BCa} for BCa interval.}

\item{show.intercept}{plot estimate and confidence interval of the intercept.}

\item{...}{other arguments.}
}
\description{
Plot confidence intervals of cluster bootstrap
}
\examples{
\dontrun{
data(opposites)
cbglm.1 <- clusbootglm(SCORE~Time*COG,data=opposites,clusterid=opposites$Subject)
plot(cbglm.1,interval.type="BCa")}
}
\author{
Mathijs Deen
}
