% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusbootglm_s3_methods.R
\name{summary.clusboot}
\alias{summary.clusboot}
\title{Summarize cluster bootstrap output}
\usage{
\method{summary}{clusboot}(object, interval.type = "BCa", ...)
}
\arguments{
\item{object}{cluster bootstrap object.}

\item{interval.type}{which confidence interval should be used. Options are \code{parametric}, \code{percentile} and \code{BCa} intervals.}

\item{...}{other arguments.}
}
\description{
Summarize cluster bootstrap output
}
\examples{
\dontrun{
data(opposites)
cbglm.1 <- clusbootglm(SCORE~Time*COG,data=opposites,clusterid=opposites$Subject)
summary(cbglm.1, interval.type="percentile")}
}
\author{
Mathijs Deen
}
