% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusbootemm.R
\name{summary.clusbootemm}
\alias{summary.clusbootemm}
\title{Summarize estimated marginal means for cluster bootstrap GLM into a grid}
\usage{
\method{summary}{clusbootemm}(object, ...)
}
\arguments{
\item{object}{object of class \code{clusbootemm}.}

\item{...}{other arguments.}
}
\description{
Returns the summary of the EMM for a \code{clusbootglm} class object.
}
\examples{
\dontrun{
medication <- medication[medication$time \%\% 1 == 0 & medication$time <=4,]
set.seed(1)
model.1 <- clusbootglm(pos~treat*time,data=medication,clusterid=id, B=5000)
emm_model.1 <- emmeans.clusbootglm(model.1, ~treat*time)
summary(emm_model.1)}
}
\author{
Mathijs Deen
}
