% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_exports.R
\name{clusbootsample}
\alias{clusbootsample}
\title{Return data for specified bootstrap sample}
\usage{
clusbootsample(object, samplenr)
}
\arguments{
\item{object}{object of class \code{clusbootglm}, created with the \code{clusbootglm} function.}

\item{samplenr}{sample number for which the data frame should be returned.}
}
\description{
Returns the full data frame for a specified bootstrap sample in a \code{clusbootglm} object.
}
\examples{
\dontrun{
data(opposites)
cbglm.1 <- clusbootglm(SCORE~Time*COG,data=opposites,clusterid=Subject)
clusbootsample(cbglm.1, samplenr=1)}
}
\author{
Mark de Rooij, Mathijs Deen
}
