% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptest.clusbootemm.R
\name{ptest.clusbootemm}
\alias{ptest.clusbootemm}
\title{Permutation test for group differences at within-subject levels on observed data}
\usage{
\method{ptest}{clusbootemm}(
  data,
  outcome,
  within,
  between,
  at.within,
  at.between,
  pn = 1000,
  progress.bar = TRUE
)
}
\arguments{
\item{data}{object of type \code{clusbootemm}}

\item{outcome}{outcome variable (i.e., the variable for which the difference should be tested).}

\item{within}{within-subject variable.}

\item{between}{between-subjects variable.}

\item{at.within}{determine for which within-subject levels (e.g., which timepoint) the difference should be tested.}

\item{at.between}{determine the groups in the difference test (should always be of length 2).}

\item{pn}{the number of permutations that should be performed.}

\item{progress.bar}{indicates whether a progress bar will be shown.}
}
\description{
The one for cluster bootstrap estimated marginal means
}
\author{
Mathijs Deen
}
