% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptest.R
\name{ptest}
\alias{ptest}
\title{Permutation tests for group differences at within-subject levels on observed data}
\usage{
ptest(data, ...)
}
\arguments{
\item{data}{data to be used for the permutation test(s).}

\item{...}{additional arguments.}
}
\description{
Generic function for performing permutation tests on either observed data or bootstrap samples
}
\author{
Mathijs Deen
}
